Added proxy profile
That reduced the mock init time to less than 40%.

Had to use a custom squid3 module because the newer squid did not accept the
old options.

It also installs the repoproxy to allow clients to use a single repository url
instead of mirrors to allow squid to properly cache them.

Change-Id: I599fd2ebaf77f6a87b247fe0f96e10d38eca2159
Signed-off-by: David Caro <dcaroest@redhat.com>