ovirt_jenkins_slave: add ovirt-guest-agent
Most jenkins slaves are VMs managed by oVirt.
Adding Guest Agent service if the slave is a vm.
There's already a 'boolean' fact to check if a machine is virtual or
not, but it does not seem to work as expected (works as a string), so
using the source value directly

Change-Id: Ib98beea08692e0f2b5439f96db019c1bf2c8a7fa
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>
Signed-off-by: David Caro <dcaroest@redhat.com>