add backup to Engine in PHX datacenter
engine-backup.sh isntalled on Engine, it runs
engine-backup and SCPs the archive to the backup server.

Backup server is configured to accept the public
key from the Engine which is saved in Hiera.

Code is mostly copied from existing gerrit backup.
Configurations are stored in hiera.

For security reasons, the keypair is generated manually
on the client, and the first connection is made by hand
to accept the backup server key.

I checked the script in a test environment and it works.

Change-Id: Ibfd79dd1940cd6df46da7c11f466cba97f91ab94