Update Jenkins puppet module version and plugin list
the new puppet module version allows control of Jenkins bundled plugins.
this will finally allow us to control all plugins via common.yaml

the updates in common.yaml reflect the current state of Jenkins plugins
installed. once this is merged we can make changes
on top of it(and be able to revert easily)

Change-Id: I7edeb92c64c7cdba8ed933d97a779ad0d6362c47
Signed-off-by: ngoldin <ngoldin@redhat.com>