rsync PHX backups off-site
This patch is intended to run on an off-site server and
will be used to copy backups from backup.phx.ovirt.org

Settings are in hiera. I tested the manifest in a lab.
Backup directory creation is not puppetized as the VM
we have is provisioned in a different way. For future
we can re-use the ovirt_backup::server class to create
partitions on servers that will do backup copying.

Change-Id: I3b6749cc3f00259392a05c16cf7eb1aa4ec78b23
Signed-off-by: Evgheni Dereveanchin <ederevea@redhat.com>