OVIRT-593 - Remove postgresql from slaves
This patch removes the following puppet class:
ovirt_jenkins_slave::engine_dao_test_runner

The class was used to install and configure postgres
on slaves which is no longer required as the job
was converted to use standard-ci

Removing the class will ensure postgres is not
installed on newly provisioned slaves.

The patch also removes the module from the Puppetfile
as this class was the only one using it.

Change-Id: I5360d821816427a6198eabbcbcf50bd7685214a7
Signed-off-by: Evgheni Dereveanchin <ederevea@redhat.com>