OVIRT-630 - fix systemd loop on jenkins slaves
This patch adds DefaultDependencies=false to scratchpad disk
service definitions. It's needed to avoid the ordering loop
as otherwise the service is implicitly started after basic.target
which is much later than we need it as we're mounting partitions.

This was tested in a test environment and it removes the loop.
I also checked that it does not affect running systems, i.e.
systemd will not re-trigger the units after they are changed
until the system is rebooted.

Change-Id: I78d0e18e2b387b9e9038d213a8930446629610d3
Signed-off-by: Evgheni Dereveanchin <ederevea@redhat.com>